templatify priority queue
